Dev => Master

+++ src/apis/course/course.resolver.ts
@@ -1,5 +1,9 @@
-import { UseGuards } from '@nestjs/common';
-import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
+import {
+  ConflictException,
+  UnauthorizedException,
+  UseGuards,
+} from '@nestjs/common';
+import { Args, Int, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { GqlAuthAccessGuard } from 'src/common/auth/gql.auth.guard';
 import { CurrentUser, ICurrentUser } from 'src/common/auth/gql.user.param';
 import { CourseService } from './course.service';
@@ -17,7 +21,7 @@ export class CourseResolver {
   }
 
   @Query(() => [Course])
-  async fetchCoursesByHost(
+  async fetchCoursesByHostId(
     @Args('hostID') hostID: string,
     @Args('page', { defaultValue: 1 }) page: number,
   ) {
@@ -25,11 +29,20 @@ export class CourseResolver {
   }
 
   @Query(() => [Course], { nullable: true })
-  async fetchCoursesSortBycreated(
+  async fetchCoursesSortByOption(
     @Args('search', { defaultValue: '' }) search: string,
     @Args('page', { defaultValue: 1 }) page: number,
+    @Args('option', { defaultValue: 'createdAt' }) option: string,
   ) {
-    return await this.courseService.searchSortByCreated(search, page);
+    if (option === 'pick') {
+      return await this.courseService.searchSortByPick(search, page);
+    } else if (option === 'discount') {
+      return await this.courseService.searchSortByDiscount(search, page);
+    } else if (option === 'createdAt') {
+      return await this.courseService.searchSortByCreated(search, page);
+    } else {
+      throw new ConflictException('옵션을 올바르게 입력해주세요.');
+    }
   }
 
   @Query(() => [Course], { nullable: true })
@@ -54,6 +67,22 @@ export class CourseResolver {
     @Args('page', { defaultValue: 1 }) page: number,
   ) {
     return await this.courseService.fetchCoursesByCategory(search, page);
+  }
+
+  @Query(() => [Course])
+  async fetchCoursesByAddress(
+    @Args('search', { defaultValue: '' }) search: string,
+    @Args('page', { defaultValue: 1 }) page: number,
+  ) {
+    return await this.courseService.searchAddress(search, page);
+  }
+
+  @Query(() => [Course])
+  async fetchCoursesByHostNickname(
+    @Args('search', { defaultValue: '' }) search: string,
+    @Args('page', { defaultValue: 1 }) page: number,
+  ) {
+    return await this.courseService.searchHostNickname(search, page);
   }
 
   @UseGuards(GqlAuthAccessGuard)
@@ -81,32 +110,56 @@ export class CourseResolver {
     return await this.courseService.cheapCourses();
   }
 
+  @Query(() => Int)
+  async howManyCourses() {
+    return await this.courseService.howManyCourses();
+  }
+
+  @UseGuards(GqlAuthAccessGuard)
+  @Query(() => Int)
+  async howManyCoursesByHost(@CurrentUser() currentUser: ICurrentUser) {
+    return await this.courseService.howManyCoursesByHost(currentUser);
+  }
+
+  @UseGuards(GqlAuthAccessGuard)
+  @Query(() => [String])
+  async myCourseRate(@CurrentUser() currentUser: ICurrentUser) {
+    return await this.courseService.myCourseRate(currentUser);
+  }
+
   @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Course)
-  createCourse(
+  async createCourse(
     @CurrentUser() currentUser: ICurrentUser,
     @Args('createCourseInput') createCourseInput: CreateCourseInput,
   ) {
-    return this.courseService.create({ createCourseInput, currentUser });
+    return await this.courseService.create({ createCourseInput, currentUser });
   }
 
-  // 호스트 본인이 쓴 글만 수정할 수 있도록 해야함
-  @Mutation(() => Course) //호스트가 바꾸는 거
+  @UseGuards(GqlAuthAccessGuard)
+  @Mutation(() => Course)
   async updateCourse(
     @Args('courseId') courseId: string,
+    @CurrentUser() currentUser: ICurrentUser,
     @Args('updateCourseInput') updateCourseInput: UpdateCourseInput,
   ) {
-    return await this.courseService.update({
-      courseId,
-      updateCourseInput,
-    });
+    const myCourse = await this.courseService.findOne({ courseId });
+    if (myCourse.host.id !== currentUser.id) {
+      throw new UnauthorizedException('호스트가 등록한 코스가 아닙니다.');
+    }
+    return await this.courseService.update(courseId, updateCourseInput);
   }
 
-  // 호스트 본인이 쓴 글만 삭제할 수 있도록 해야함
+  @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Boolean)
-  deleteCourse(
-    @Args('courseId') courseId: string, //
+  async deleteCourse(
+    @CurrentUser() currentUser: ICurrentUser,
+    @Args('courseId') courseId: string,
   ) {
+    const myCourse = await this.courseService.findOne({ courseId });
+    if (myCourse.host.id !== currentUser.id) {
+      throw new UnauthorizedException('호스트가 등록한 코스가 아닙니다.');
+    }
     return this.courseService.delete({ courseId });
   }
 }
